+++ app/src/js/components/Granules/overview.js
@@ -68,6 +68,7 @@ class GranulesOverview extends React.Component {
       workflowMeta: defaultWorkflowMeta,
       selected: [],
       isInfixSearch: false,
+      isArchivedSearch: false,
     };
   }
 
@@ -88,7 +89,7 @@ class GranulesOverview extends React.Component {
 
   generateQuery() {
     const { queryParams } = this.props;
-    return { ...queryParams };
+    return { ...queryParams, archived: this.state.isArchivedSearch };
   }
 
   generateBulkActions() {
@@ -121,6 +122,10 @@ class GranulesOverview extends React.Component {
 
   setIsInfixSearch = (value) => {
     this.setState({ isInfixSearch: value });
+  };
+
+  setIsArchivedSearch = (value) => {
+    this.setState({ isArchivedSearch: value });
   };
 
   applyWorkflow(granuleId) {
@@ -160,7 +165,7 @@ class GranulesOverview extends React.Component {
     const { dropdowns } = collections;
     const { dropdowns: providerDropdowns } = providers;
     const { count, queriedAt } = list.meta;
-    const { isInfixSearch } = this.state;
+    const { isInfixSearch, isArchivedSearch } = this.state;
 
     return (
       <div className="page__component">
@@ -212,7 +217,7 @@ class GranulesOverview extends React.Component {
               placeholder="Granule ID"
               searchKey="granules"
               infixBoolean={isInfixSearch}
-              archived="false"
+              archived={isArchivedSearch}
             />
             <ListFilters>
               <Checkbox
@@ -223,6 +228,15 @@ class GranulesOverview extends React.Component {
                 inputLabel="Infix"
                 className="infix-search"
                 tip="Toggle between prefix and infix search. When enabled, the search field matches substrings instead of prefixes."
+              />
+              <Checkbox
+                id="chk_isArchivedSearch"
+                checked={isArchivedSearch}
+                onChange={this.setIsArchivedSearch}
+                label="Include"
+                inputLabel="Archived"
+                className="archived-search"
+                tip="Toggle inclusion of archived records in search results"
               />
               <Dropdown
                 options={statusOptions}